+++ i18n/text.go
@@ -5,6 +5,11 @@ type Text struct {
 	dic map[string]string
 }
 
+func (a *Text) String() string {
+	lang := GetLanguage()
+	return a.Get(lang)
+}
+
 func (a *Text) Get(lang string) string {
 	s, ok := a.dic[lang]
 	if !ok {