+++ main/main.go
@@ -13,12 +13,6 @@ import (
 	"github.com/aliyun/aliyun-cli/i18n"
 )
 
-var usage = `
-	Alibaba Cloud CLI(Command Line Interface)
-	Usage:
-`
-
-
 /**
 ## Configure
 
@@ -33,10 +27,11 @@ $ aliyuncli configure
 	$ aliyuncli Ecs StartInstance --InstanceId your_instance_id
 	$ aliyuncli Rds DescribeDBInstances
 
-## 用HTTPS(SSL/TLS)通信
+## use HTTPS(SSL/TLS)
 
 	$ aliyuncli Ecs DescribeInstances --secure
 */
+const Version = "0.50 BETA"
 
 var profileName string
 var library = meta.LoadLibrary(resource.NewReader())
@@ -44,10 +39,19 @@ var helper = openapi.NewHelper(library)
 var configureCommand = config.NewConfigureCommand()
 
 func main() {
+	//
+	// set language from current configuration
+	profile, err := config.LoadCurrentProfile()
+	if err != nil {
+		cli.Errorf("get current failed %s", err)
+		return
+	}
+	i18n.SetLanguage(profile.Language)
+
 	rootCmd := &cli.Command{
 		Name: "aliyun",
-		Short: i18n.T("Alibaba Cloud Command Line Interface Version 0.33 Beta", "阿里云CLI命令行工具 0.33 Beta"),
-		Usage: "aliyun <product> <operation> --parameter1 value1 --parameter2 value2 ...",
+		Short: i18n.T("Alibaba Cloud Command Line Interface Version " + Version, "阿里云CLI命令行工具 " + Version),
+		Usage: "aliyun <product> <operation> [--parameter1 value1 --parameter2 value2 ...]",
 		Sample: "aliyun ecs DescribeRegions",
 		EnableUnknownFlag: true,
 		Run: func(ctx *cli.Context, args []string) error {
@@ -59,22 +63,30 @@ func main() {
 	}
 
 	rootCmd.Flags().StringVar(&profileName, "profile", "default",
-		i18n.En("root.profile", "use configured profile"))
+		i18n.T("use --profile <profileName> to use specified profile", "使用 --profile <profileName> 来使用指定的配置"))
+
+	rootCmd.Flags().Add(cli.Flag{Name: "secure", Assignable:false,
+		Usage: i18n.T("use --secure to force https", "使用 --secure 开关强制使用https方式调用")})
 
 	rootCmd.Flags().Add(cli.Flag{Name: "force", Assignable:false,
-		Usage: i18n.T("root.force", "call OpenAPI without check")})
+		Usage: i18n.T("use --force to skip api and parameters check", "添加 --force 开关可跳过API与参数的合法性检查")})
+
 	rootCmd.Flags().Add(cli.Flag{Name: "endpoint", Assignable:true,
-		Usage: i18n.En("root.endpoint", "use assigned endpoint")})
+		Usage: i18n.T("use --endpoint <endpoint> to assign endpoint", "使用 --endpoint <endpoint> 来指定接入点地址")})
+
 	rootCmd.Flags().Add(cli.Flag{Name: "region", Assignable:true,
-		Usage: i18n.En("root.region","use assigned region")})
+		Usage: i18n.T("use --region <regionId> to assign region", "使用 --region <regionId> 来指定访问地域")})
 	rootCmd.Flags().Add(cli.Flag{Name: "version", Assignable:true,
-		Usage: i18n.En("root.version", "assign product version")})
+		Usage: i18n.T("use --version <YYYY-MM-DD> to assign product api version", "使用 --version <YYYY-MM-DD> 来指定访问的API版本")})
+
 	rootCmd.Flags().Add(cli.Flag{Name: "header", Assignable:true, Repeatable:true,
-		Usage: i18n.En("root.header", "add custom HTTP header with --header x-foo=bar")})
+		Usage: i18n.T("use --header X-foo=bar to add custom HTTP header, repeatable", "使用 --header X-foo=bar 来添加特定的HTTP头, 可多次添加")})
+
 	rootCmd.Flags().Add(cli.Flag{Name: "body", Assignable:true,
-		Usage: i18n.En("root.body", "assign http body in Restful call")})
+		Usage: i18n.T("use --body $(cat foo.json) to assign http body in RESTful call", "使用 --body $(cat foo.json) 来指定在RESTful调用中的HTTP包体")})
+
 	rootCmd.Flags().Add(cli.Flag{Name: "body-file", Assignable:true, Hidden: true,
-		Usage: i18n.En("root.body", "assign http body in Restful call with local file")})
+		Usage: i18n.T("assign http body in Restful call with local file", "")})
 
 	rootCmd.AddSubCommand(configureCommand)
 	rootCmd.AddSubCommand(command.NewTestCommand())